building loader improvements
