+++ project/carpg.vcxproj.filters
@@ -58,6 +58,12 @@
     <Filter Include="2. engine\gui\dialogs">
       <UniqueIdentifier>{22378397-713d-49a2-bddf-1773b059fcbc}</UniqueIdentifier>
     </Filter>
+    <Filter Include="3. game\system\objects">
+      <UniqueIdentifier>{40df2bfe-e01b-47b1-9ab8-c6bcab9c75aa}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="3. game\system\buildings">
+      <UniqueIdentifier>{832797a9-e416-46ec-bdc8-965ed5a4bd43}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="source\resource\PickLanguage.cpp">
@@ -420,9 +426,6 @@
     <ClCompile Include="source\game\Init.cpp">
       <Filter>3. game</Filter>
     </ClCompile>
-    <ClCompile Include="source\game\system\Building.cpp">
-      <Filter>3. game\system</Filter>
-    </ClCompile>
     <ClCompile Include="source\game\GameStats.cpp">
       <Filter>3. game</Filter>
     </ClCompile>
@@ -632,13 +635,25 @@
       <Filter>1. core</Filter>
     </ClCompile>
     <ClCompile Include="source\game\system\BaseObject.cpp">
-      <Filter>3. game\system</Filter>
-    </ClCompile>
-    <ClCompile Include="source\game\system\ObjectLoader.cpp">
-      <Filter>3. game\system</Filter>
+      <Filter>3. game\system\objects</Filter>
     </ClCompile>
     <ClCompile Include="source\game\system\BaseUsable.cpp">
-      <Filter>3. game\system</Filter>
+      <Filter>3. game\system\objects</Filter>
+    </ClCompile>
+    <ClCompile Include="source\game\system\ObjectLoader.cpp">
+      <Filter>3. game\system\objects</Filter>
+    </ClCompile>
+    <ClCompile Include="source\game\system\BuildingLoader.cpp">
+      <Filter>3. game\system\buildings</Filter>
+    </ClCompile>
+    <ClCompile Include="source\game\system\Building.cpp">
+      <Filter>3. game\system\buildings</Filter>
+    </ClCompile>
+    <ClCompile Include="source\game\system\BuildingScript.cpp">
+      <Filter>3. game\system\buildings</Filter>
+    </ClCompile>
+    <ClCompile Include="source\game\system\BuildingGroup.cpp">
+      <Filter>3. game\system\buildings</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1008,9 +1023,6 @@
     <ClInclude Include="source\core\FastFunc.h">
       <Filter>1. core</Filter>
     </ClInclude>
-    <ClInclude Include="source\game\system\Building.h">
-      <Filter>3. game\system</Filter>
-    </ClInclude>
     <ClInclude Include="source\game\system\Content.h">
       <Filter>3. game\system</Filter>
     </ClInclude>
@@ -1271,10 +1283,19 @@
       <Filter>1. core</Filter>
     </ClInclude>
     <ClInclude Include="source\game\system\BaseObject.h">
-      <Filter>3. game\system</Filter>
+      <Filter>3. game\system\objects</Filter>
     </ClInclude>
     <ClInclude Include="source\game\system\BaseUsable.h">
-      <Filter>3. game\system</Filter>
+      <Filter>3. game\system\objects</Filter>
+    </ClInclude>
+    <ClInclude Include="source\game\system\BuildingGroup.h">
+      <Filter>3. game\system\buildings</Filter>
+    </ClInclude>
+    <ClInclude Include="source\game\system\Building.h">
+      <Filter>3. game\system\buildings</Filter>
+    </ClInclude>
+    <ClInclude Include="source\game\system\BuildingScript.h">
+      <Filter>3. game\system\buildings</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
