+++ project/source/core/Text.h
@@ -24,6 +24,39 @@ struct AnyString
 	cstring s;
 };
 
+inline bool operator == (const string& s1, const AnyString& s2)
+{
+	return s1 == s2.s;
+}
+inline bool operator == (const AnyString& s1, const string& s2)
+{
+	return s2 == s1.s;
+}
+inline bool operator == (cstring s1, const AnyString& s2)
+{
+	return strcmp(s1, s2.s) == 0;
+}
+inline bool operator == (const AnyString& s1, cstring s2)
+{
+	return strcmp(s1.s, s2) == 0;
+}
+inline bool operator != (const string& s1, const AnyString& s2)
+{
+	return s1 != s2.s;
+}
+inline bool operator != (const AnyString& s1, const string& s2)
+{
+	return s2 != s1.s;
+}
+inline bool operator != (cstring s1, const AnyString& s2)
+{
+	return strcmp(s1, s2.s) != 0;
+}
+inline bool operator != (const AnyString& s1, cstring s2)
+{
+	return strcmp(s1.s, s2) != 0;
+}
+
 //-----------------------------------------------------------------------------
 struct AnyStringNull
 {
