+++ project/source/game/Tournament.cpp
@@ -48,7 +48,7 @@ bool Game::IfUnitJoinTournament(Unit& u)
 //=================================================================================================
 void Game::GenerateTournamentUnits()
 {
-	Vec3 pos = city_ctx->FindBuilding(content::BG_ARENA)->walk_pt;
+	Vec3 pos = city_ctx->FindBuilding(BuildingGroup::BG_ARENA)->walk_pt;
 	tournament_master = FindUnitByIdLocal("arena_master");
 
 	// przenieœ herosów przed arenê
@@ -100,7 +100,7 @@ void Game::UpdateTournament(float dt)
 			tournament_timer += dt;
 
 		// do³¹czanie cz³onków dru¿yny
-		const Vec3& walk_pt = city_ctx->FindBuilding(content::BG_ARENA)->walk_pt;
+		const Vec3& walk_pt = city_ctx->FindBuilding(BuildingGroup::BG_ARENA)->walk_pt;
 		for(Unit* unit : Team.members)
 		{
 			if(unit->busy == Unit::Busy_No && Vec3::Distance2d(unit->pos, tournament_master->pos) <= 16.f && !unit->dont_attack && IfUnitJoinTournament(*unit))
@@ -145,7 +145,7 @@ void Game::UpdateTournament(float dt)
 					}
 				}
 
-				city_ctx->FindInsideBuilding(content::BG_ARENA, tournament_arena);
+				city_ctx->FindInsideBuilding(BuildingGroup::BG_ARENA, tournament_arena);
 				tournament_state2 = 3;
 				tournament_round = 0;
 				tournament_master->busy = Unit::Busy_Yes;
@@ -595,7 +595,7 @@ void Game::TournamentTrain(Unit& u)
 //=================================================================================================
 void Game::CleanArena()
 {
-	InsideBuilding* arena = city_ctx->FindInsideBuilding(content::BG_ARENA);
+	InsideBuilding* arena = city_ctx->FindInsideBuilding(BuildingGroup::BG_ARENA);
 
 	// wyrzuæ ludzi z areny
 	for(vector<Unit*>::iterator it = at_arena.begin(), end = at_arena.end(); it != end; ++it)
