+++ project/source/game/location/InsideBuilding.cpp
@@ -6,6 +6,7 @@
 #include "SaveState.h"
 #include "Content.h"
 #include "Door.h"
+#include "BuildingGroup.h"
 
 //=================================================================================================
 InsideBuilding::~InsideBuilding()
@@ -141,13 +142,13 @@ void InsideBuilding::Load(HANDLE file, bool local)
 	if(LOAD_VERSION >= V_0_5)
 	{
 		ReadString1(file);
-		type = content::FindBuilding(BUF);
+		type = Building::Get(BUF);
 	}
 	else
 	{
 		OLD_BUILDING old_type;
 		ReadFile(file, &old_type, sizeof(old_type), &tmp, nullptr);
-		type = content::FindOldBuilding(old_type);
+		type = Building::GetOld(old_type);
 	}
 	assert(type != nullptr);
 	ReadFile(file, &level_shift, sizeof(level_shift), &tmp, nullptr);
@@ -262,7 +263,7 @@ void InsideBuilding::Load(HANDLE file, bool local)
 	}
 
 	// konwersja krzese³ w sto³ki
-	if(LOAD_VERSION < V_0_2_12 && type->group == content::BG_INN)
+	if(LOAD_VERSION < V_0_2_12 && type->group == BuildingGroup::BG_INN)
 	{
 		auto chair = BaseUsable::Get("chair"),
 			stool = BaseUsable::Get("stool");
