+++ project/source/game/system/BaseUsable.cpp
@@ -2,13 +2,13 @@
 #include "Core.h"
 #include "BaseUsable.h"
 
+//-----------------------------------------------------------------------------
 vector<BaseUsable*> BaseUsable::usables;
 
-BaseUsable* BaseUsable::TryGet(cstring id)
+//=================================================================================================
+BaseUsable* BaseUsable::TryGet(const AnyString& id)
 {
-	assert(id);
-
-	for(BaseUsable* use : usables)
+	for(auto use : usables)
 	{
 		if(use->id == id)
 			return use;
