+++ project/source/game/system/ContentLoader.h
@@ -1,5 +1,9 @@
 #pragma once
 
+//-----------------------------------------------------------------------------
+#include "Content.h"
+
+//-----------------------------------------------------------------------------
 class ContentLoader
 {
 public:
